mediacapture-streams/MediaStreamTrackEvent-constructor.https.html

<!doctype html>
<title>MediaStreamTrackEvent constructor</title>
<link rel="help" href="https://w3c.github.io/mediacapture-main/#mediastreamtrackevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

test(() => {
  assert_equals(MediaStreamTrackEvent.length, 2);
  assert_throws(new TypeError, () => {
    new MediaStreamTrackEvent("type");
  });
  assert_throws(new TypeError, () => {
    new MediaStreamTrackEvent("type", null);
  });
  assert_throws(new TypeError, () => {
    new MediaStreamTrackEvent("type", undefined);
  });
}, "The eventInitDict argument is required");

test(() => {
  assert_throws(new TypeError, () => {
    new MediaStreamTrackEvent("type", {});
  });
  assert_throws(new TypeError, () => {
    new MediaStreamTrackEvent("type", { track: null });
  });
  assert_throws(new TypeError, () => {
    new MediaStreamTrackEvent("type", { track: undefined });
  });
}, "The eventInitDict's track member is required.");

test(() => {
  // a MediaStreamTrack instance is needed to test, any instance will do.
  const context = new AudioContext();
  const dest = context.createMediaStreamDestination();
  const track = dest.stream.getTracks()[0];
  assert_true(track instanceof MediaStreamTrack);
  const event = new MediaStreamTrackEvent("type", { track });
  assert_equals(event.type, "type");
  assert_equals(event.track, track);
}, "The MediaStreamTrackEvent instance's track attribute is set.");

</script>
